# contracts/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "소비나침반 API",
    "description": "동성로 한식 프로토타입 API와 승인 기반 Gemini 조건 추출 준비",
    "version": "3.2.0"
  },
  "paths": {
...
    "/api/meta": {
      "get": {
        "summary": "Meta",
        "operationId": "meta_api_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaResponse"
                }
              }
            }
          },
          "500": {
            "description": "처리 중 서버 오류가 발생했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "활성 스냅샷 또는 의존 서비스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/parse-query": {
      "post": {
        "summary": "Parse Query",
        "operationId": "parse_query_api_parse_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseResponse"
                }
              }
            }
          },
          "413": {
            "description": "요청 본문이 정책상 허용된 크기를 초과했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 형식 또는 값이 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "처리 중 서버 오류가 발생했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "활성 스냅샷 또는 의존 서비스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/recommendations": {
      "post": {
        "summary": "Recommendations",
        "operationId": "recommendations_api_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "409": {
            "description": "스냅샷이 현재 결과와 일치하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 형식 또는 값이 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "처리 중 서버 오류가 발생했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "활성 스냅샷 또는 의존 서비스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/places/{place_id}": {
      "get": {
        "summary": "Place",
        "operationId": "place_api_places__place_id__get",
        "parameters": [
          {
            "name": "place_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Place Id"
            }
          },
          {
            "name": "snapshot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponse"
                }
              }
            }
          },
          "404": {
            "description": "요청한 리소스를 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "스냅샷이 현재 결과와 일치하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 형식 또는 값이 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "처리 중 서버 오류가 발생했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "활성 스냅샷 또는 의존 서비스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/budget/estimate": {
      "post": {
        "summary": "Budget Estimate",
        "operationId": "budget_estimate_api_budget_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetEstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            }
          },
          "404": {
            "description": "요청한 리소스를 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "스냅샷이 현재 결과와 일치하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "요청 본문이 정책상 허용된 크기를 초과했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 형식 또는 값이 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "처리 중 서버 오류가 발생했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "활성 스냅샷 또는 의존 서비스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BudgetEstimateRequest": {
        "properties": {
          "conditions": {
            "$ref": "#/components/schemas/Conditions"
          },
          "snapshot_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Snapshot Id"
          },
          "place_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Place Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "conditions",
          "snapshot_id",
          "place_ids"
        ],
        "title": "BudgetEstimateRequest"
      },
...
      "ErrorResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "MetaResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "minLength": 1,
            "title": "Request Id"
          },
          "snapshot_id": {
            "type": "string",
            "minLength": 1,
            "title": "Snapshot Id"
          },
          "policy_version": {
            "type": "string",
            "minLength": 1,
            "title": "Policy Version"
          },
          "data_mode": {
            "type": "string",
            "minLength": 1,
            "title": "Data Mode"
          },
          "reference_now": {
            "type": "string",
            "minLength": 1,
            "title": "Reference Now"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "regions": {
            "items": {
              "type": "string",
              "enum": [
                "dongseongro",
                "suseongmot"
              ]
            },
            "type": "array",
            "title": "Regions"
          },
          "categories": {
            "items": {
              "type": "string",
              "enum": [
                "restaurant",
                "cafe"
              ]
            },
            "type": "array",
            "title": "Categories"
          },
          "atmosphere_tags": {
            "items": {
              "type": "string",
              "enum": [
                "quiet",
                "conversation",
                "lively",
                "solo_friendly"
              ]
            },
            "type": "array",
            "title": "Atmosphere Tags"
          },
          "parser_mode": {
            "type": "string",
            "title": "Parser Mode"
          },
          "explanation_mode": {
            "type": "string",
            "title": "Explanation Mode"
          },
          "map_mode": {
            "type": "string",
            "title": "Map Mode"
          },
          "llm": {
            "$ref": "#/components/schemas/AdapterStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "snapshot_id",
          "policy_version",
          "data_mode",
          "reference_now",
          "warnings",
          "regions",
          "categories",
          "atmosphere_tags",
          "parser_mode",
          "explanation_mode",
          "map_mode",
          "llm"
        ],
        "title": "MetaResponse"
      },
...
      "ParseQueryRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500,
            "title": "Text"
          },
          "external_ai_consent": {
            "type": "boolean",
            "title": "External Ai Consent",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ParseQueryRequest"
      },
      "ParseResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "minLength": 1,
            "title": "Request Id"
          },
          "snapshot_id": {
            "type": "string",
            "minLength": 1,
            "title": "Snapshot Id"
          },
          "policy_version": {
            "type": "string",
            "minLength": 1,
            "title": "Policy Version"
          },
          "data_mode": {
            "type": "string",
            "minLength": 1,
            "title": "Data Mode"
          },
          "reference_now": {
            "type": "string",
            "minLength": 1,
            "title": "Reference Now"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "parser_mode": {
            "type": "string",
            "title": "Parser Mode"
          },
          "draft": {
            "$ref": "#/components/schemas/ParseDraft"
          },
          "field_meta": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldMeta"
            },
            "type": "object",
            "title": "Field Meta"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "confirmation_required": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Confirmation Required"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "snapshot_id",
          "policy_version",
          "data_mode",
          "reference_now",
          "warnings",
          "parser_mode",
          "draft",
          "field_meta",
          "missing_fields",
          "confirmation_required"
        ],
        "title": "ParseResponse"
      },
...
      "PlaceResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "minLength": 1,
            "title": "Request Id"
          },
          "snapshot_id": {
            "type": "string",
            "minLength": 1,
            "title": "Snapshot Id"
          },
          "policy_version": {
            "type": "string",
            "minLength": 1,
            "title": "Policy Version"
          },
          "data_mode": {
            "type": "string",
            "minLength": 1,
            "title": "Data Mode"
          },
          "reference_now": {
            "type": "string",
            "minLength": 1,
            "title": "Reference Now"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "place": {
            "$ref": "#/components/schemas/PlaceData"
          },
          "price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PriceData"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceData"
            },
            "type": "array",
            "title": "Evidence"
          },
          "area_timeseries": {
            "items": {
              "$ref": "#/components/schemas/AreaMetricData"
            },
            "type": "array",
            "title": "Area Timeseries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "snapshot_id",
          "policy_version",
          "data_mode",
          "reference_now",
          "warnings",
          "place",
          "price",
          "evidence",
          "area_timeseries"
        ],
        "title": "PlaceResponse"
      },
...
      "RecommendationRequest": {
        "properties": {
          "conditions": {
            "$ref": "#/components/schemas/Conditions"
          },
          "snapshot_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Snapshot Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "conditions",
          "snapshot_id"
        ],
        "title": "RecommendationRequest"
      },
      "RecommendationResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "minLength": 1,
            "title": "Request Id"
          },
          "snapshot_id": {
            "type": "string",
            "minLength": 1,
            "title": "Snapshot Id"
          },
          "policy_version": {
            "type": "string",
            "minLength": 1,
            "title": "Policy Version"
          },
          "data_mode": {
            "type": "string",
            "minLength": 1,
            "title": "Data Mode"
          },
          "reference_now": {
            "type": "string",
            "minLength": 1,
            "title": "Reference Now"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItem"
            },
            "type": "array",
            "title": "Items"
          },
          "insufficient_data_items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItem"
            },
            "type": "array",
            "title": "Insufficient Data Items"
          },
          "excluded_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Excluded Counts"
          },
          "active_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Features"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "snapshot_id",
          "policy_version",
          "data_mode",
          "reference_now",
          "warnings",
          "items",
          "insufficient_data_items",
          "excluded_counts",
          "active_features"
        ],
        "title": "RecommendationResponse"
      }
    }
  }
}